node struct
