+++ libcentrifugo/node/api_test.go
@@ -416,64 +416,40 @@ type TestCommand struct {
 	Token     string `json:"token"`
 }
 
-func UnmarshalExplicit(b []byte) (*TestCommand, error) {
-	var cmd TestCommand
-	err := json.Unmarshal(b, &cmd)
-	if err != nil {
-		return nil, err
+func testArguments(user, timestamp, info, token string) error {
+	if user != "1" {
+		panic("err")
 	}
-	return &cmd, nil
+	return nil
 }
 
-func UnmarshalInterface(method string, b []byte) (interface{}, error) {
-	var cmd TestCommand
-	err := json.Unmarshal(b, &cmd)
-	if err != nil {
-		return nil, err
+func testStruct(cmd *TestCommand) error {
+	if cmd.User != "1" {
+		panic("err")
 	}
-	return &cmd, nil
+	return nil
 }
 
-func BenchmarkUnmarshalExplicit(b *testing.B) {
-	cmd := TestCommand{
-		User:      "1",
-		Timestamp: "21212122",
-		Info:      "wyf84ft638f7h3487f483f3478f6h348f68h437hf63478f",
-		Token:     "465f7gtyd45f6tgs54d6fsd5rf6t",
-	}
-	cmdBytes, err := json.Marshal(cmd)
-	if err != nil {
-		panic(err)
-	}
-	b.ResetTimer()
+func BenchmarkPassArguments(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		cmd, err := UnmarshalExplicit(cmdBytes)
+		err := testArguments("1", "21212122", "wyf84ft638f7h3487f483f3478f6h348f68h437hf63478f", "2dw223d23f23f")
 		if err != nil {
 			panic(err)
 		}
-		if cmd.User != "1" {
-			panic("111")
-		}
 	}
 }
 
-func BenchmarkUnmarshalInterface(b *testing.B) {
-	cmd := TestCommand{
-		User:      "1",
-		Timestamp: "21212122",
-		Info:      "wyf84ft638f7h3487f483f3478f6h348f68h437hf63478f",
-		Token:     "465f7gtyd45f6tgs54d6fsd5rf6t",
-	}
-	cmdBytes, _ := json.Marshal(cmd)
-	b.ResetTimer()
+func BenchmarkPassStruct(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		cmdInt, err := UnmarshalInterface("connect", cmdBytes)
+		cmd := TestCommand{
+			User:      "1",
+			Timestamp: "21212122",
+			Info:      "wyf84ft638f7h3487f483f3478f6h348f68h437hf63478f",
+			Token:     "2dw223d23f23f",
+		}
+		err := testStruct()
 		if err != nil {
 			panic(err)
-		}
-		cmd := cmdInt.(*TestCommand)
-		if cmd.User != "1" {
-			panic("111")
 		}
 	}
 }
